delete reader with tests, check request for node env or ip address

+++ routes/hardDelete.js
@@ -10,29 +10,39 @@ const { Note } = require('../models/Note')
 const { Notebook } = require('../models/Notebook')
 const { NoteContext } = require('../models/NoteContext')
 const { Tag } = require('../models/Tag')
+const { Reader } = require('../models/Reader')
 
 const timestamp = new Date(Date.now() - 86400 * 1000).toISOString()
 
 module.exports = function (app) {
   app.use('/', router)
   router.route('/hardDelete').delete(async function (req, res, next) {
-    await Publication.query()
-      .delete()
-      .where('deleted', '<', timestamp)
-    await Note.query()
-      .delete()
-      .where('deleted', '<', timestamp)
-    await Notebook.query()
-      .delete()
-      .where('deleted', '<', timestamp)
-    await NoteContext.query()
-      .delete()
-      .where('deleted', '<', timestamp)
+    if (
+      process.env.NODE_ENV === 'test' ||
+      req.connection.remoteAddress === '10.0.0.1'
+    ) {
+      await Publication.query()
+        .delete()
+        .where('deleted', '<', timestamp)
+      await Note.query()
+        .delete()
+        .where('deleted', '<', timestamp)
+      await Notebook.query()
+        .delete()
+        .where('deleted', '<', timestamp)
+      await NoteContext.query()
+        .delete()
+        .where('deleted', '<', timestamp)
+      await Tag.query()
+        .delete()
+        .where('deleted', '<', timestamp)
+      await Reader.query()
+        .delete()
+        .where('deleted', '<', timestamp)
 
-    await Tag.query()
-      .delete()
-      .where('deleted', '<', timestamp)
-
-    res.status(204).end()
+      res.status(204).end()
+    } else {
+      res.status(403)
+    }
   })
 }
